always render library into sidebar & refactor components
